Move ProgressListener interface to common ui package.
