+++ src/main/java/org/mastodon/revised/mamut/feature/DefaultMamutFeatureComputerService.java
@@ -13,11 +13,11 @@ import org.mastodon.graph.algorithm.TopologicalSort;
 import org.mastodon.graph.object.ObjectEdge;
 import org.mastodon.graph.object.ObjectGraph;
 import org.mastodon.graph.object.ObjectVertex;
-import org.mastodon.revised.mamut.ProgressListener;
 import org.mastodon.revised.model.feature.Feature;
 import org.mastodon.revised.model.feature.FeatureComputer;
 import org.mastodon.revised.model.feature.FeatureComputerService;
 import org.mastodon.revised.model.mamut.Model;
+import org.mastodon.revised.ui.ProgressListener;
 import org.scijava.InstantiableException;
 import org.scijava.app.StatusService;
 import org.scijava.log.LogService;
